fix(elasticsearch): Solved merge conflicts and circular imports after branch update
